% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_dag.r
\name{assert_dag}
\alias{assert_dag}
\title{assert that a network is a DAG}
\usage{
assert_dag(network)
}
\arguments{
\item{networtk}{igraph instance}
}
\description{
assert that a network is a DAG
}
\note{
if dag has cicles, raises an Exception

TODO: dovrebbe tornare una lista di cicli se trovati. con la nuova
versione di igraph non funziona.
}
