% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{build_connection}
\alias{build_connection}
\title{Factory di connessioni al database}
\usage{
build_connection(env = getenv())
}
\arguments{
\item{env}{environment label}
}
\description{
Factory di connessioni al database
}
\note{
Funzione interna
}
