% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{buildConnection}
\alias{buildConnection}
\title{Factory di connessioni al database}
\usage{
buildConnection(env = getenv(), con = NULL)
}
\description{
Factory di connessioni al database
}
\note{
Funzione interna
}
