% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dag.r
\name{check_dag}
\alias{check_dag}
\title{Controlla se il network e' un DAG}
\usage{
check_dag(network)
}
\description{
TODO: dovrebbe tornare una lista di cicli se trovati. con la nuova versione di igraph non funziona.
}
