% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDAG.r
\name{checkDAG}
\alias{checkDAG}
\title{Controlla se il network e' un DAG}
\usage{
checkDAG(network)
}
\description{
TODO: dovrebbe tornare una lista di cicli se trovati. con la nuova versione di igraph non funziona.
}
