% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{clutter_with_params_and_return}
\alias{clutter_with_params_and_return}
\title{questa funzione orla la formula del grafo come una funzione}
\usage{
clutter_with_params_and_return(f, name, deps, funcName)
}
\arguments{
\item{f}{function task to be converted as function}

\item{name}{task name}

\item{deps}{array di dipendenze}

\item{funcName}{nome della funzione definita}
}
\value{
Ritorna un character array `is.character(ret) == TRUE`
}
\description{
I parametri della funzione ritornata sono le dipendenze della serie,
ed aggiunge il nome della funzione al termine per dichiarare il
dato ritornato
}
