% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{.clutter_with_params}
\alias{.clutter_with_params}
\title{questa funzione orla la formula del grafo come una funzione}
\usage{
.clutter_with_params(f, name, deps)
}
\arguments{
\item{f}{function task to be converted as function}

\item{name}{task name}

\item{deps}{character array di dipendenze}
}
\value{
Ritorna una una funzione `is.character(ret) == TRUE`
}
\description{
I parametri della funzione ritornata sono le dipendenze della serie
}
