% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_data_frame}
\alias{convert_data_frame}
\title{Converte un dataframe in timeseries}
\usage{
convert_data_frame(df)
}
\arguments{
\item{df}{data.frame con le serie storiche}
}
\value{
a List
}
\description{
Viene usato nella conversione dei dati dal db
}

