% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.r
\name{countRolling}
\alias{countRolling}
\title{conta le versioni rolling del grafo con tag `tag`}
\usage{
countRolling(x)
}
\arguments{
\item{x}{istanza di grafo}

\item{con}{connessione al DB}
}
\value{
un intero ad indicare il numero di versioni rolling salvate sul DB
}
\description{
conta le versioni rolling del grafo con tag `tag`
}
