% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.r
\name{.createGraph}
\alias{.createGraph}
\title{crea ex-novo un istanza di grafo nel databae}
\usage{
.createGraph(g, tag)
}
\arguments{
\item{x}{istanza di Grafo}

\item{tag}{identificativo della versione}

\item{con}{connessione al DB}
}
\description{
crea ex-novo un istanza di grafo nel databae
}
