% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{db_settings}
\alias{db_settings}
\title{Reads db connections settings}
\usage{
db_settings()

db_settings(flush)
}
\arguments{
\item{flush}{if `TRUE` removes any option saved in the current
session and reloads the settings}
}
\value{
a list containing the infos used to connect via DBI
}
\description{
Tryes :
  - `$HOME/.GrafoDB/GrafoDB.ini` file
  - Internal settings.
}
