% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{dbDisconnect}
\alias{dbDisconnect}
\title{overrides dbDisconnect to do nothings in case pgConnect is active in options}
\usage{
dbDisconnect(con)
}
\arguments{
\item{con}{connection to be closed}
}
\description{
overrides dbDisconnect to do nothings in case pgConnect is active in options
}

