% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{dbSettings}
\alias{dbSettings}
\title{Reads db connections settings}
\usage{
dbSettings()
}
\arguments{
\item{flush}{if `TRUE` removes any option recored in the current
session and reloads the settings}
}
\value{
a list containing the infos used to connect via DBI/RPostgreSQL
}
\description{
Tryes :
  - `$HOME/.GrafoDB/GrafoDB.ini` file
  - Internal settings.
}
