% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{.declutter_function}
\alias{.declutter_function}
\title{funzione per eliminare le definizione 'function' dalle formule per il GrafoDB}
\usage{
.declutter_function(f)
}
\arguments{
\item{f}{formula in formato testo}
}
\description{
funzione per eliminare le definizione 'function' dalle formule per il GrafoDB
}
