% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_graph.r
\name{delete_graph}
\alias{delete_graph}
\title{Deletes a Graph from the DB}
\usage{
delete_graph(tag)
}
\arguments{
\item{tag}{`tag` as the primary key of the Graph}
}
\description{
Deletes a graph by its tag
}
