% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadati.r
\name{deleteMeta_impl}
\alias{deleteMeta_impl}
\title{delete un metadato dal DBI}
\usage{
deleteMeta_impl(x, name, key, value)
}
\arguments{
\item{x}{istanza di grafo}

\item{name}{nome della serie da cui eliminare il metadato}

\item{key}{nome del metadato}

\item{value}{valore del metadato (se non specificato, rimuove
tutti i  metadati
          con la chiave specificata)}
}
\description{
delete un metadato dal DBI
}
