% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{deleteMeta}
\alias{deleteMeta}
\title{Removes an attribute}
\usage{
deleteMeta(object, ts_name, attr_name, attr_value)
}
\arguments{
\item{object}{graph instance}

\item{ts_name}{name of the timeseries}

\item{attr_name}{name of the attribute}

\item{attr_value}{value of the attributes}
}
\description{
Removes an attribute
}
