% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{deps}
\alias{deps}
\title{Yields the dependencies of an object}
\usage{
deps(x, name)
}
\arguments{
\item{x}{istanza di GrafoDB}

\item{name}{nome della serie}
}
\value{
a character array of object names, NULL if not present
}
\description{
Yields the dependencies of an object
}
