\docType{methods}
\name{describe}
\alias{describe}
\title{Funzioni del package `grafo`}
\usage{
describe(graph, nodes, order, mode, plot)
}
\arguments{
  \item{object}{L'istanza di grafo che si vuole
  visualizzare}

  \item{mode}{Analisi in ingresso ('in') o uscita ('out')}

  \item{plot}{TRUE: plot del grafo, FALSE: lista dei nodi
  in area di lavoro}
}
\description{
implementazione di describe di `package::grafo`
}
\seealso{
`grafo::describe`
}

