% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.r
\name{diff.GrafoDB}
\alias{diff.GrafoDB}
\title{diff implementation for GrafoDB}
\usage{
\method{diff}{GrafoDB}(x, ...)
}
\arguments{
\item{x}{GrafoDB instance}

\item{`...`}{eventual GrafoDB instance; if not `stop`}
}
\description{
Returns diffs as data and formulas
}
