% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{disconnect}
\alias{disconnect}
\title{proxy the dbDisconnect based on `getenv` value}
\usage{
disconnect(con, env = getenv())
}
\arguments{
\item{con}{connection to disconnect}

\item{env}{environment to check against}
}
\description{
proxy the dbDisconnect based on `getenv` value
}
