% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{dollar_impl}
\alias{dollar_impl}
\title{internal function to implement $<-}
\usage{
dollar_impl(x, name, value)
}
\arguments{
\item{x}{GrafoDB instance}

\item{name}{key of the object to be set}

\item{value}{value of the object to set}
}
\value{
an invisible `x`
}
\description{
internal function to implement $<-
}
