% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edit.GrafoDB}
\alias{edit.GrafoDB}
\title{edita una formula del `GrafoDB`}
\usage{
edit(g, nome)
}
\arguments{
\item{g}{istanza di grafo}

\item{name}{nome della serie da editare}
}
\description{
E' un metodo di tipo S3 per essere compliant con la definizione di `edit`
}

