% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edita.r
\name{.edita}
\alias{.edita}
\title{edit a GrafoDB formula}
\usage{
.edita(x, name, ...)
}
\arguments{
\item{x}{GrafoDB instance}

\item{name}{name of the object to be edited}

\item{`...`}{other eventual options}
}
\description{
edit a GrafoDB formula
}
