% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{edita}
\alias{edita}
\alias{edita,GrafoDB,character-method}
\alias{GrafoDB,character}
\title{Edita un la formula di una serie storica.}
\usage{
edita(x, name)

\S4method{edita}{GrafoDB,character}(x, name, ...)
}
\arguments{
\item{x}{istanza di grafo}

\item{name}{nome della serie storica}
}
\value{
il grafo con la formula modificata
}
\description{
Edita un la formula di una serie storica.
}
