% Generated by roxygen2 (4.0.2): do not edit by hand
\name{elimina}
\alias{elimina}
\title{Elimina un edizione del grafo}
\usage{
elimina(tag)
}
\arguments{
\item{tag}{`tag` che distingue in modo univoco il grafo ed i suoi dati}
}
\description{
Cancella dal database un edizione del grafo partendo dal suo `tag`
}

