% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{evaluate_plain}
\alias{evaluate_plain}
\title{Patch to evaluate}
\usage{
evaluate_plain(x, ids = names(x))
}
\arguments{
\item{x}{GrafoDB instance}

\item{ids}{object names to be evaluated}
}
\value{
GrafoDB with ids evaluated
}
\description{
Patch to evaluate
}
\note{
this is useful to evaluate the topological sort
  without multicore
}
