% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{evaluate_single}
\alias{evaluate_single}
\title{Evaluates a single object function identified with `name`}
\usage{
evaluate_single(name, graph)
}
\arguments{
\item{name}{`character` nome della serie}

\item{graph}{istanza di `GrafoDB`}
}
\value{
la serie storica calcolata.
}
\description{
Evaluates a single object function identified with `name`
}
