% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{.evaluateSingle}
\alias{.evaluateSingle}
\alias{evaluateSingle}
\title{Valuta un singolo oggetto del grafo identificato da `name`}
\usage{
.evaluateSingle(name, object)

evaluateSingle(name, object)
}
\arguments{
\item{name}{`character` nome della serie}

\item{graph}{istanza di `GrafoDB`}
}
\value{
la serie storica calcolata.
}
\description{
Valuta un singolo oggetto del grafo identificato da `name`
}
\note{
la scelta di evaluateSingle ricade su .evaluateSingle1, le altre due
      implementazioni NON SUPPORTANO LE SERIE ELEMENTARI
}
