% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{exponent}
\alias{exponent}
\title{returns the exponent of a floating point numeber (eng representation)}
\usage{
exponent(x)
}
\value{
the exponent of the eng. representation of the number
}
\description{
returns the exponent of a floating point numeber (eng representation)
}
