% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.r
\name{expr_impl}
\alias{expr_impl}
\title{Ritorna le formule del GrafoDB}
\usage{
expr_impl(x, nomi, echo = FALSE)
}
\arguments{
\item{x}{istanza di GrafoDB}

\item{nomi}{array di nomi di serie storiche}

\item{echo}{if TRUE prints the formula on on the standard output}
}
\description{
Ritorna (o mostra su console, stdout) le formule delle serie `nomi`
}
