% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{expr}
\alias{expr}
\alias{expr,GrafoDB,character-method}
\title{Formule del `GrafoDB`}
\usage{
expr(x, nomi)

\S4method{expr}{GrafoDB,character}(x, nomi, echo = FALSE)
}
\arguments{
\item{x}{istanza di oggetto R}

\item{nomi}{character array di nomi di serie storiche}

\item{echo}{stampa con un messaggio su standard output il
valore della formula}
}
\value{
`list` con nomi (i nomi sono gli stess del parametro `nomi`)
        con le formule
}
\description{
Ritorna come named list le formule per ogni serie specificata in `nomi`
}
\examples{
\dontrun{
g <- GrafoDB(...)
expr(g, "TETSZ0AC") # ritorna list(TETSZ0AC = "TETSZ0AC = ASTSZ0AC...")
}
}
