% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_deps.r
\name{find_dep_from_error}
\alias{find_dep_from_error}
\title{this function returns the name of the missing object in
evaluation}
\usage{
find_dep_from_error(cond = NULL)
}
\arguments{
\item{cond}{error obtained in tryCatch}
}
\value{
single name of missing object blamed in `cond`, NA if not found
}
\description{
this function returns the name of the missing object in
evaluation
}
\seealso{
find_deps
}
