% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_deps.r
\name{find_deps}
\alias{find_deps}
\title{finds and returns the deps of a formula as a character vector}
\usage{
find_deps(g, formula)
}
\arguments{
\item{g}{anything that has names end data as
list (env, list, dataset, GrafoDB)}

\item{formula}{the formula itself}
}
\value{
list of names not defined in `formula` and needed to evaluate it
}
\description{
Given a formula for `C` like
   C <- A+B
   D <- C+B/2
the function returns `A` and `B` but not `C`
}
