% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.r
\name{fixConflicts}
\alias{fixConflicts}
\title{contrassegna come risolti i conflitti salvati sul DB}
\usage{
fixConflicts(x)
}
\arguments{
\item{x}{istanza di GrafoDB}

\item{name}{nome della serie da cui eliminare un conflitto}
}
\description{
E' appannaggio dell'utente riusolvere i conflitti nella sua sessione
e provvedere a salvate un Grafo consistente.
}
