% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formule.r
\name{formule}
\alias{formule}
\title{This programs shows all the editions of a formula}
\usage{
formule(g, name)
}
\arguments{
\item{g}{GrafoDB instance}

\item{name}{name of the object to be inspected}
}
\value{
a data.frame with unique formulas and `tag`
}
\description{
This programs shows all the editions of a formula
}
