% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadati.r
\name{get_metadata_impl}
\alias{get_metadata_impl}
\title{Ritorna i metadati della serie `name` in `x`}
\usage{
get_metadata_impl(x, name)
}
\arguments{
\item{x}{istanza di grafo}

\item{name}{nome della serie storica}
}
\value{
data.frame contenente i metadati della serie
}
\description{
Ritorna i metadati della serie `name` in `x`
}
