% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.r
\name{getChangedSeries}
\alias{getChangedSeries}
\title{Trova le serie che sono cambiate nel database}
\usage{
getChangedSeries(x)
}
\arguments{
\item{x}{istanza di Grafo}

\item{con}{eventuale connessione al database (se non presente, ne crea una)}
}
\value{
lista di nomi di serie cambiate sul database rispetto ad X
}
\description{
Trova le serie che sono cambiate nel database
}
\note{
funzione interna
}
