% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.r
\name{getConflicts}
\alias{getConflicts}
\alias{getConflicts,GrafoDB-method}
\title{Ritorna i conflitti per un Grafo}
\usage{
getConflicts(x, name = NULL, con = NULL)

\S4method{getConflicts}{GrafoDB}(x, name = NULL, con = NULL)
}
\arguments{
\item{x}{istanza del GrafoDB}

\item{name}{vettore di nomi di serie}

\item{con}{connection to the DB, if `NULL` the connections 
gets created and closed in the method}
}
\value{
un data.frame con le informazioni del conflitto
}
\description{
Ritorna i conflitti di un grafo, tutti per serie specificate dal
parametro `name`
}
