\docType{methods}
\name{getData}
\alias{getData}
\title{Funzioni del package `grafo`}
\usage{
getData(graph, tsNames)
}
\arguments{
  \item{graph}{un oggetto di classe('grafo')}

  \item{tsNames}{nome di una serie o lista di nomi di
  serie}
}
\description{
implementazione di getData di `package::grafo`
}
\seealso{
`grafo::getData`
}

