% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{.getdata}
\alias{.getdata}
\title{Ottiene i dati dal GrafoDB}
\usage{
.getdata(x, i)
}
\arguments{
\item{x}{istanza di `GrafoDB`}

\item{i}{character array di nomi di serie storiche}
}
\value{
ritorna una named list con all'interno le serie storiche. Se l'array e'
        di un solo elemento, ritorna direttamente la serie storica
        (questo e' un side-effect, non mi piace)
}
\description{
I dati possono provenire direttamente dal Database se non modificati nella sessione
corrente; altriumenti vengono restituiti i dati che l'utente ha appena modificato
ma non ancora reso persistenti
}
\note{
se i e' un singolo nome e non esiste nel DB, la funzione termina con errore
}
