% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{getdb}
\alias{getdb}
\title{Carica i dati dal DB}
\usage{
getdb(x, name)
}
\arguments{
\item{x}{istance di GrafoDB}

\item{name}{nome serie}
}
\value{
una serie o una lista di serie
}
\description{
Carica i dati direttamente dal DB senza necessita' d'inizializzare
un `GrafoDB`
}
