\docType{methods}
\name{getDependencies}
\alias{getDependencies}
\title{Funzioni del package `grafo`}
\usage{
getDependencies(graph, tsName)
}
\arguments{
  \item{object}{un oggetto di classe('grafo')}

  \item{tsName}{nome di una serie}
}
\description{
implementazione di getDependencies di `package::grafo`
}
\seealso{
`grafo::getDependencies`
}

