% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{getenv}
\alias{getenv}
\title{Returns the enviroment for GrafoDB}
\usage{
getenv()
}
\value{
the value of the GRAFODB_ENV, if not specified, defaults
   to `prod`
}
\description{
Returns the value of environment variable `GRAFODB_ENV`
}
