% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{getMeta}
\alias{getMeta}
\alias{getMeta,GrafoDB,character,character-method}
\alias{getMeta,GrafoDB,character,ANY-method}
\alias{getMeta,GrafoDB,missing,missing-method}
\title{Ottiene i valori di un `metadato` per una `serie``}
\usage{
getMeta(x, serie, metadato)

getMeta(x, serie)

\S4method{getMeta}{GrafoDB,character,character}(x, serie, metadato)

\S4method{getMeta}{GrafoDB,character,ANY}(x, serie, metadato)

\S4method{getMeta}{GrafoDB,missing,missing}(x)
}
\arguments{
\item{x}{istanza di `GrafoDB`}

\item{serie}{nome della serie d'interesse}

\item{metadato}{nome del metadato di cui si vogliono i valori}
}
\value{
un character vector con i nomi dei valori del metadati
}
\description{
Ritorna i valori di un metadato (o piu' metadati) specificati
dal parametro `metadato` della serie specificata dal
parametro `serie`
}
\examples{
\dontrun{
  g <- GrafoDB()
  getMeta(g, "TETSZ0AC",
             "TAVOLA_DI_OUTPUT") ## ritorna tutti i valori del metadato
                                 ## 'TAVOLA_DI_OUTPUT' per 'TETSZ0AC'
}
}
