\docType{methods}
\name{getTask}
\alias{getTask}
\title{Funzioni del package `grafo`}
\usage{
getTask(graph, tsName)
}
\arguments{
  \item{object}{un oggetto di classe('grafo')}

  \item{tsName}{nome di una serie}
}
\description{
implementazione di getTask di `package::grafo`
}
\seealso{
`grafo::getTask`
}

