% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\docType{class}
\name{GrafoDB-class}
\alias{GrafoDB-class}
\alias{GrafoDB}
\title{Main GrafoDB class}
\description{
Class for manipulating data described with formulas and evaluated as
a topological sorting of nodes in a graph data structure.
This class provides a persistence layer towards generic DBMS
(by now only PostgreSQL) has been implemented
}
\section{Slots}{

\describe{
\item{\code{tag}}{edition of this Graph}

\item{\code{network}}{\link[igraph]{igraph} containing the DAG}

\item{\code{data}}{\link[hash]{hash} containing changed data of this \link{GrafoDB}}

\item{\code{functions}}{\link[hash]{hash} containing function data
of this \link{GrafoDB}}

\item{\code{ordinal}}{ordinale dei dati storici (0 per la produzione corrente)}

\item{\code{timestamp}}{of the current GrafoDB}

\item{\code{touched}}{object being modified}

\item{\code{edges}}{the edges being modified}

\item{\code{dbdati}}{data.frame containing data}

\item{\code{dbformule}}{data.frame containing functions}

\item{\code{helper}}{SQLHelper class}
}}

\examples{
\dontrun{
   g = GrafoDB("cf10") # istanzia il grafo chiamato 'cf10'
                       # in questo caso ordinal e' 0
   g = GrafoDB("cf10p2") # istanzia il grafo cf10 con provvisorio p2;
                         # in questo caso ordinal e' 2
}
}
