% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r, R/subsetting.r
\name{$,GrafoDB-method}
\alias{$,GrafoDB-method}
\alias{GrafoDB,character}
\alias{$<-,GrafoDB,ANY-method}
\alias{GrafoDB,ANY}
\alias{[,GrafoDB,character,missing,ANY-method}
\alias{GrafoDB,character,missing,ANY}
\alias{[[,GrafoDB,character,missing-method}
\alias{[<-,GrafoDB,character,missing,ANY-method}
\alias{[[<-,GrafoDB,character,missing,ANY-method}
\title{Extract data from GrafoDB}
\usage{
\S4method{$}{GrafoDB}(x, name)

\S4method{$}{GrafoDB,ANY}(x, name) <- value

\S4method{[}{GrafoDB,character,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{GrafoDB,character,missing}(x, i, j, ...)

\S4method{[}{GrafoDB,character,missing,ANY}(x, i, j, ...) <- value

\S4method{[[}{GrafoDB,character,missing,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{an instance of GrafoDB}

\item{name}{name of the object}

\item{value}{the value to be set, can be anything}

\item{i}{name to be set}

\item{j}{unused (ignored)}

\item{...}{nothing important (just to comply on generic)}

\item{drop}{I dunno why this is here :)}
}
\value{
a Dataset or a single objec

a Dataset or a single object

a Dataset or a single object

a list or a single object
}
\description{
returns data from GrafoDB (seen here as a key-value container)
by specifying the key (a string) for a value

returns data from GrafoDB (seen here as a key-value container)
by specifying the key (a string) for a value

returns data from GrafoDB (seen here as a key-value container)
by specifying the key (a string) for a value

returns data from GrafoDB (seen here as a key-value container)
by specifying the key (a string) for a value
}
