% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{grafodb <-}
\alias{grafodb <-}
\alias{grafodb}
\title{Alias constructor for GrafoDB}
\usage{
grafodb(...)
}
\arguments{
\item{...}{params passed to GrafoDB}
}
\value{
instance of GrafoDB
}
\description{
Alias constructor for GrafoDB
}
\seealso{
GrafoDB::GrafoDB
}
