\docType{methods}
\name{initialize}
\alias{GrafoDB-initialize}
\alias{initialize}
\title{costruttore per la classe GrafoDB}
\usage{
\S4method{initialize}{GrafoDB}(.Object, tag = "cf10")
}
\arguments{
  \item{.Object}{ An object: see the Details section.}
}
\description{
costruttore per la classe GrafoDB
}

