% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{init_grafo_impl}
\alias{init_grafo_impl}
\title{Funzione d'inizializzazione del grafo.}
\usage{
init_grafo_impl(object, tag = "cf10", con = NULL)
}
\arguments{
\item{object}{(creato da new)}

\item{tag}{tag del grafo (default=`cf10`)}

\item{con}{Connessione la DB}
}
\value{
un istanza di grafo popolata correttamente secono i parametri (`tag`)
}
\description{
Questa funzione va utilizzata nell'initialize S4 dell'oggetto `GrafoDB`
}
\note{
e' stata scorporata dall'initialize S4 per finalita' di debug
}
