% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlhelper.r
\name{init_sql_helper}
\alias{init_sql_helper}
\title{initializes the SQL helper}
\usage{
init_sql_helper(object)

init_sql_helper(object, path)

init_sql_helper(object, path, type)
}
\arguments{
\item{object}{the SQLHelper instance}

\item{type}{the kind of SQLhelper to initialize.
Depends on environment (`getenv`) if not passwd by}
}
\value{
the object initialized
}
\description{
initializes the SQL helper
}
