% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_deps.r
\name{isError}
\alias{isError}
\title{True if param is an error}
\usage{
isError(cond)
}
\arguments{
\item{cond}{argument to check}
}
\value{
`TRUE` if cond is an error
}
\description{
True if param is an error
}
