\docType{methods}
\name{isPrimitive}
\alias{isPrimitive}
\title{Funzioni del package `grafo`}
\usage{
isPrimitive(graph, tsName)
}
\arguments{
  \item{graph}{un oggetto di classe('grafo')}

  \item{tsName}{nome della serie di cui si vuole conoscere
  se e' Primitiva}
}
\description{
implementazione di isPrimitive di `package::grafo`
}
\seealso{
`grafo::isPrimitive`
}

