% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadati.r
\name{.keys}
\alias{.keys}
\title{ritorna le chiavi di un metadato}
\usage{
.keys(x)
}
\arguments{
\item{x}{istanza di grafo}
}
\value{
ritorna un dataframe con le chiavi dei metadati
}
\description{
ritorna le chiavi di un metadato
}

