% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{.leaves}
\alias{.leaves}
\title{Ritorna le figlie del GrafoDB}
\usage{
.leaves(x)
}
\arguments{
\item{x}{Grafo}
}
\value{
la lista delle foglie del grafo
}
\description{
Ritona tutti i nodi del grafo le che non hanno archi uscenti
}
