% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{leaves}
\alias{leaves}
\title{Ritorna le foglie del grafo}
\usage{
leaves(x)
}
\arguments{
\item{x}{istanza di grafo}
}
\value{
la lista di nomi delle foglie del grafo
}
\description{
Ritorna le foglie del grafo
}
