% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{listAggregates}
\alias{listAggregates}
\title{Funzioni del package `grafo`}
\usage{
listAggregates(graph)
}
\description{
implementazione di listAggregates di `package::grafo`
}
\seealso{
`grafo::listAggregates`
}
