% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{listElementaries}
\alias{listElementaries}
\alias{listElementaries,GrafoDB-method}
\title{Funzioni del package `grafo`}
\usage{
listElementaries(graph)

\S4method{listElementaries}{GrafoDB}(graph)
}
\arguments{
\item{graph}{GrafoDB instance}
}
\description{
implementazione di listElementaries di `package::grafo`
}
\seealso{
`grafo::listElementaries`
}
