% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{listNodes}
\alias{listNodes}
\alias{listNodes,GrafoDB-method}
\title{Funzioni del package `grafo`}
\usage{
listNodes(graph)

\S4method{listNodes}{GrafoDB}(graph)
}
\arguments{
\item{graph}{GrafoDB instance}
}
\description{
implementazione di listNodes di `package::grafo`
}
\seealso{
`grafo::listNodes`
}
