% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence_utils.r
\name{load_table}
\alias{load_table}
\title{Legge tutta la tabella dei dati per un determinato tag}
\usage{
load_table(table_name, tag, con = NULL)
}
\arguments{
\item{table_name}{nome della tabella}

\item{tag}{nome del tag}

\item{con}{Connessione al DB}
}
\value{
un data.frame con i dati delle serie storiche
}
\description{
Legge tutta la tabella dei dati per un determinato tag
}
\note{
funzione interna
}
