% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence_utils.r
\name{loadTable}
\alias{loadTable}
\title{Legge tutta la tabella dei dati per un determinato tag}
\usage{
loadTable(table_name, tag)

loadTable(table_name, tag, con)
}
\arguments{
\item{table_name}{nome della tabella}

\item{tag}{nome del tag}
}
\value{
un data.frame con i dati delle serie storiche
}
\description{
Legge tutta la tabella dei dati per un determinato tag
}
\note{
funzione interna
}
