% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.r
\name{.lookup_formula}
\alias{.lookup_formula}
\title{Ricerca nelle formule del grafo}
\usage{
.lookup_formula(x, key)
}
\arguments{
\item{x}{istanza di grafo}

\item{key}{Stringa da cercare}
}
\value{
lista di nomi che hanno `key` nella formula
}
\description{
Ricerca nelle formule del grafo
}
