% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.r
\name{.lookup}
\alias{.lookup}
\title{Ricerca nei metadati del grafo}
\usage{
.lookup(x, key, value)
}
\arguments{
\item{x}{istanza di grafo}

\item{key}{chiave del metadato}

\item{value}{valore del metadato}
}
\value{
lista di nomi che matchano la condizione `key==valy`
}
\description{
Ricerca nei metadati del grafo
}
