% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.r
\name{lookup_dati_impl}
\alias{lookup_dati_impl}
\title{Ricerca nelle formule del grafo}
\usage{
lookup_dati_impl(x, key)
}
\arguments{
\item{x}{istanza di grafo}

\item{key}{numero da cercare}
}
\value{
lista di nomi che matchano `key` nella formula
}
\description{
Ricerca nelle formule del grafo
}
