% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{names}
\alias{names}
\title{Returns all the keys in this graph.}
\usage{
\S4method{names}{GrafoDB}(x)
}
\arguments{
\item{x}{GrafoDB instance}
}
\value{
all the nodes in the graph, topologically sorted.
}
\description{
The array returned is already sorted by topological sorting
algorithm
}
