% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{navigate}
\alias{navigate}
\alias{navigate,GrafoDB-method}
\title{Funzioni del package `grafo`}
\usage{
navigate(object, nodes = NULL, order = 1L, mode = "out")

\S4method{navigate}{GrafoDB}(object, nodes = NULL, order = 1L, mode = "out")
}
\arguments{
\item{object}{GrafoDB instance}

\item{nodes}{array of node names}

\item{order}{number o levels (orders) to extend}

\item{mode}{direction of navigation: `"in"``, `"out"`` or `"all"`,
default `"out"``}
}
\description{
'navigates' in the graph
}
\details{
This method lets you move through the nodes of the graph data
structure, movigin surfing the edges of the graph.
}
