% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.r
\name{next_rolling_name}
\alias{next_rolling_name}
\title{Costruice il progressivo per il grafo `x`}
\usage{
next_rolling_name(x, con)
}
\arguments{
\item{x}{GrafoDB instance}

\item{con}{DB connection}
}
\description{
Costruice il progressivo per il grafo `x`
}
