% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.r
\name{patch}
\alias{patch}
\title{apply a patch (from a diff)}
\usage{
patch(x, diff_, column)
}
\arguments{
\item{x}{GrafoDB instance}

\item{diff_}{differences (evaluated with diff.GrafoDB)}

\item{column}{which column of diff_ to apply on Grafo}
}
\description{
a diff on GrafoDB are differences in formulas between two editions of a
GrafoDB.
}
\note{
Experimental
}
\seealso{
diff.GrafoDB
}
