% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{pgConnect}
\alias{pgConnect}
\title{trying to behave like a connection pool
(with a single connection :( )}
\usage{
pgConnect()
}
\value{
a Connection to Postgresql
}
\description{
trying to behave like a connection pool
(with a single connection :( )
}
\note{
this stores the connection into options and retrieves it back
}
