% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.r
\name{readBinary}
\alias{readBinary}
\title{Legge un GrafoDB dal filesystem in formato binario}
\usage{
readBinary(path)
}
\arguments{
\item{path}{percorso del file con il GrafoDB}
}
\value{
GrafoDB contenuto nel file `path`
}
\description{
Legge un GrafoDB dal filesystem in formato binario
}
