% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_archi.r
\name{rebuild_edges}
\alias{rebuild_edges}
\title{this function extract deps based on pattern matching on nodes}
\usage{
rebuild_edges(x)
}
\description{
this function extract deps based on pattern matching on nodes
}
