% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.r
\name{rename_impl}
\alias{rename_impl}
\title{Rinomina una serie del grafo}
\usage{
rename_impl(x, vecchio, nuovo)
}
\arguments{
\item{x}{istanza di `GrafoDB`}

\item{vecchio}{nome vecchio da sostituire}

\item{nuovo}{nome nuovo da sostituire}
}
\value{
grafo modificato
}
\description{
L'operazione lavora direttamente sui dati in modo persistente.
}
