% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{rename}
\alias{rename}
\title{Rinomina una serie del grafo}
\usage{
rename(x, vecchio, nuovo)
}
\arguments{
\item{x}{istanza di `GrafoDB`}

\item{vecchio}{nome vecchio da sostituire}

\item{nuovo}{nome nuovo da sostituire}
}
\value{
grafo modificato
}
\description{
L'operazione lavora direttamente sui dati in modo persistente.
}
\examples{
\dontrun{
    g <- GrafoDB("test")
    rename(g, "TS1", "TS2")
    "TS2" \%in\% names(g) # questo e' `TRUE`
    "TS1" \%in\% names(g) # questo e' `FALSE`
}
}
