% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{rilasci}
\alias{rilasci}
\title{Ritorna la lista dei rilasci presenti nel database}
\usage{
rilasci()
}
\arguments{
\item{filtro}{filtro da applicare alla ricerca sul tag del grafo}
}
\value{
data.frame con tutti i rilasci
}
\description{
Ritorna la lista dei rilasci presenti nel database
}
\examples{
\dontrun{
   rilasci() # ritorna tutti i rilasci
   rilasci("cf") # ritorna tutti i rilasci con contenenti cf nel tag
}
}
