% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNode.r
\name{rm_node_impl}
\alias{rm_node_impl}
\title{Elimina un nodo dal `GrafoDB`}
\usage{
rm_node_impl(graph, ts_name, recursive = FALSE)
}
\arguments{
\item{graph}{istanza di `GrafoDB`}

\item{ts_name}{nomi di serie da eliminare}

\item{recursive}{`TRUE` se l'eliminazione deve essere rivorsiva sugli archi
uscenti di ogni serie nel parametro `ts_name`. `FALSE` altrimenti. Se il
parametro e' impostato a `FALSE` e' condizione necessaria che le serie in
`ts_name` siano tutte foglie, ovvero serie senza archi uscenti}
}
\description{
L'eliminazione prevede l'eliminazione dai dati, formule, archi e metadati
}
\note{
Metodo interno
}
\seealso{
rmNode
}
