% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{rmNode}
\alias{rmNode}
\title{Elimina un nodo dal `GrafoDB`}
\usage{
rmNode(graph, tsName, recursive)
}
\arguments{
\item{graph}{istanza di `GrafoDB`}

\item{tsName}{nomi di serie da eliminare}

\item{recursive}{`TRUE` se l'eliminazione deve essere rivorsiva sugli archi
uscenti di ogni serie nel parametro `tsName`.
`FALSE` altrimenti. Se il parametro e' impostato a `FALSE` e'
condizione necessaria che le serie in `tsName` siano tutte
foglie, ovvero serie senza archi uscenti}
}
\value{
il grafo modificato
}
\description{
L'eliminazione prevede l'eliminazione dai dati, formule, archi e metadati
}
\note{
Metodo interno
}
\seealso{
.rmNode
}
