% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{.roots}
\alias{.roots}
\title{Ritorna le radici del GrafoDB}
\usage{
.roots(x)
}
\arguments{
\item{x}{Grafo}
}
\value{
la lista delle radici del grafo
}
\description{
Ritona tutti i nodi del grafo le che non hanno archi entranti
}

