% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{roots}
\alias{roots}
\alias{roots,GrafoDB-method}
\title{Returns the roots of the Graph}
\usage{
roots(x)

\S4method{roots}{GrafoDB}(x)
}
\arguments{
\item{x}{GrafoDB instance}
}
\value{
character array with the root names
}
\description{
'Roots' are those nodes *not* having incoming edges
}
