% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.r
\name{saveBinary}
\alias{saveBinary}
\title{Salva un istanza di grafo sul file system}
\usage{
saveBinary(x, path)
}
\arguments{
\item{x}{istanza del GrafoDB}

\item{path}{percorso del file su cui salvare il grafo}
}
\description{
Salva un istanza di grafo sul file system
}
\note{
il restore si fa con il comando `readBinary`
}
