% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.r
\name{save_graph_impl}
\alias{save_graph_impl}
\title{Funzione per salvare un grafo}
\usage{
save_graph_impl(x, tag)

save_graph_impl(x)
}
\description{
La funzione controlla la presenza di eventuali conflitti e necessita'
di risincronizzare i dati del DB con quelli presenti nel Grafo.
}
\details{
\itemize{
  \item{"1"}{
     Identificare le serie aggregate (solo formule) - primitive (solo dati)
     cambiate, escludendo eventuali conflitti}
  \item{"2"}{Caricarle nel grafo}
  \item{"3"}{Rieseguirle}
  \item{"4"}{Risalvare il grafo}
}

La funzione controlla se esistono conflitti nel seguente modo:
\itemize{
  \item{"dati"}{
       Se esistono serie primitive nel DB e nel grafo
       in sessione che sono state aggiornate in
       contemporanea}
  \item{"formule"}{
       Se esistono formule nel DB e nel grafo in
       sessione aggiornati in contemporanea}
}

Qualora uno dei due casi si verificasse il grafo va in "conflitto",
vengono salvate sia le proprie modifiche che le modifiche fatte da
altri e si attende la risoluzione del conflitto attraverso i metodi
`fixConflict`. La soluzione dei conflitti non e' un atto di fede:
occorre incontrarsi e decidere quale "formula" o quale versione dei dati
sia da preferire.
}
\note{
\url{https://osiride-public.utenze.bankit.it/group/894smf/trac/cfin/ticket/31849}
}
\seealso{
saveGraph
}
