% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{saveGraph}
\alias{saveGraph}
\title{implementazione di saveGraph per `GrafoDB` dal package `grafo`}
\usage{
saveGraph(object)
}
\arguments{
\item{object}{istanza di `GrafoDB`}

\item{path}{erroneamente, dovuta al generic su `grafo` questo sarebbe il "tag" da
dare al grafo. Non c'e' modo di ovviare questo problema. Vedere il
prototipo di funzione di `.saveGraph`, e' sicuramente piu' chiaro.}
}
\description{
La funzione si preoccupa di serializzare la transazione al DB, e controllare
l'eventualita' di Conflitti, segnalandoli all'utenza.
}
\note{
R sometimes sucks.
}
\seealso{
.saveGraph
}

