% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{searchNode}
\alias{searchNode}
\title{Funzioni del package `grafo`}
\usage{
searchNode(graph, tsNames)
}
\description{
implementazione di searchNode di `package::grafo`
}
\seealso{
`grafo::searchNode`
}

