% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{ser}
\alias{ser}
\title{Valuta una singola serie del grafo e ritorna il risultato}
\usage{
ser(x, name)
}
\arguments{
\item{x}{un istanza di GrafoDB}

\item{name}{nome della serie da valutare}

\item{debug}{debug mode (`FALSE` default)}
}
\value{
la serie valutata
}
\description{
Valuta una singola serie del grafo e ritorna il risultato
}
