% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{setupdb}
\alias{setupdb}
\title{Setup the DB}
\usage{
setupdb()
}
\arguments{
\item{overwrite}{if a config file exists, setting this
to `FALSE` raises an Exception}
}
\description{
Interactively reads db parameters and stores into your home dir
}

