% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{should_create_schema}
\alias{should_create_schema}
\title{Returns true if the app should create the DB schema}
\usage{
should_create_schema(con)
}
\arguments{
\item{con}{DB connection}
}
\value{
`TRUE` if db schema creation is needed, `FALSE` otherwise
}
\description{
Returns true if the app should create the DB schema
}
