% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\docType{methods}
\name{show}
\alias{show}
\alias{show,GrafoDB-method}
\title{prints info about object}
\usage{
\S4method{show}{GrafoDB}(object)
}
\arguments{
\item{object}{object to be printed-out}
}
\description{
prints info about object
}
