% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{show}
\alias{show}
\alias{show,GrafoDB}
\title{Prints a brief summary for the graph}
\usage{
show(object)
}
\arguments{
\item{object}{GrafoDB instance}
}
\description{
Prints a brief summary for the graph
}
\note{
this is the default method called in the REPL when
  an object is printed out
}
