% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.r
\name{showHistory}
\alias{showHistory}
\title{Mostra le release precedenti (provvisorio) del grafo}
\usage{
showHistory(x)
}
\arguments{
\item{x}{oggetto R su cui mostrare la history. Puo' essere un istanza
di `GrafoDB` (da cui si prende la tag con `slot(x, 'tag')`), oppure
la `tag` stessa}
}
\value{
un dataframe con le informazioni (autore, last_updated, ordinale)
}
\description{
La funzione dato il tag, mostra la history di un `GrafoDB`
Ovvero tutte le release che hanno sovrascritto la release principale
Quella senza p
}

