% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{showInternalChanges}
\alias{showInternalChanges}
\title{Funzioni del package `grafo`}
\usage{
showInternalChanges(graph)
}
\description{
implementazione di showInternalChanges di `package::grafo`
}
\seealso{
`grafo::showInternalChanges`
}

