% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlhelper.r
\docType{methods}
\name{sql_by_key}
\alias{sql_by_key}
\alias{sql_by_key,SQLHelper,character-method}
\alias{SQLHelper,character,ANY-method}
\title{Factory di query (stringhe) SQL.}
\usage{
sql_by_key(x, .key, ...)

\S4method{sql_by_key}{SQLHelper,character}(x, .key, ...)
}
\arguments{
\item{x}{istanza di SQLHelper}

\item{.key}{nome della query nel file INI}
}
\value{
un character array contenente la query SQL
}
\description{
Ritorna le query SQL cosi' come storicizzate nell'SQLHelper. Sostituisce
eventuali parametri (key=value) della query nell'ellipsis del metodo
}
\details{
Le query hanno una forma "select * from --tab-- where param='--value--'";
Il metodo si occupa di sostiture '--tab--' e '--value--' secondo i valori
passati come argomento:
Es: `sql_by_key(helper, KEYQUERY, tab="A", value="B")`
risultera' in `select * from A where param='B'`
}
