% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlhelper.r
\docType{class}
\name{SQLHelper-class}
\alias{SQLHelper-class}
\alias{SQLHelper}
\title{Questa classe provvede a disaccoppiare SQL dal codice dell'app}
\usage{
SQLHelper(...)
}
\description{
Separa le query SQL con delle chiavi, riponendo il linguaggio SQL in dei file
editabili e accedendo alle singole query con delle chiavi.
}
\details{
Le query vengono inizializzate in un file .ini
}
\section{Slots}{

\describe{
\item{\code{sqlContainer}}{a list of key-values for the SQL queries}

\item{\code{type}}{the type of syntax}
}}

