% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlhelper.r
\docType{class}
\name{SQLHelper}
\alias{SQLHelper}
\title{SQLHelper}
\usage{
SQLHelper(path)
}
\arguments{
\item{path}{path al file contenente le query (default a file interno al
package GrafoDB)}

\item{type}{tipo di database supportati (per ora "PostgreSQL" e "SQLite").
Quest'opzione e' sovrascritta in caso esistano delle options di
ambiente chiamate "SQLHelperType"}
}
\description{
Questa classe provvede a disaccoppiare SQL dal codice dell'app
}
\details{
Separa le query SQL con delle chiavi, riponendo il linguaggio SQL in dei file
editabili e accedendo alle singole query con delle chiavi.

Le query vengono inizializzate in un file .ini
}
