% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{sqlite_connect}
\alias{sqlite_connect}
\title{builds a SQLite connection}
\usage{
sqlite_connect()
}
\value{
a connection to SQLite
}
\description{
builds a SQLite connection
}
\note{
this is used for testing, the connection is to an in memory instance
}
