\name{.subsetting}
\alias{.subsetting}
\title{implementazione per GrafoDB del subsetting '[<-'}
\usage{
.subsetting(x, i, j, ..., value)
}
\arguments{
  \item{x}{istanza di `GrafoDB`}

  \item{i}{un character}

  \item{j}{un'altro character o `missing`}

  \item{...}{ellipsis dio solo sa a cosa servono}

  \item{value}{valore settato}
}
\description{
implementazione per GrafoDB del subsetting '[<-'
}
\note{
funzione interna
}

