% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickets.r
\name{ticket}
\alias{ticket}
\title{ritorna i tickets aperti nel trac}
\usage{
ticket(num)
}
\arguments{
\item{id}{id del ticket}
}
\value{
una rappresentazione a ticket (alla meglio di R)
}
\description{
ritorna i tickets aperti nel trac
}

