% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{tsdiff}
\alias{tsdiff}
\title{Checks if a TimeSeries is different from another}
\usage{
tsdiff(a, b)

tsdiff(a, b, thr)
}
\arguments{
\item{a}{timeseries}

\item{b}{timeseries}

\item{thr}{threshold for difference}
}
\value{
`TRUE` if `a`!=`b`, `FALSE` otherwise
}
\description{
It's a predicate, returns `TRUE` if:
\itemize{
\item a - b != 0
\item index(a) != index(b)
}
}
