% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadati.r
\name{.values}
\alias{.values}
\title{Ritorna i valori dei metadati (o per singolo metadato)}
\usage{
.values(x)
}
\arguments{
\item{x}{istanza di grafo}

\item{key}{chiave del metadato}
}
\value{
lista di valori per metadato
}
\description{
Ritorna i valori dei metadati (o per singolo metadato)
}
