% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadati.r
\name{values_by_key_impl}
\alias{values_by_key_impl}
\title{Ritorna i valori dei metadati (o per singolo metadato)}
\usage{
values_by_key_impl(x, key = NULL)
}
\arguments{
\item{x}{istanza di grafo}

\item{key}{chiave del metadato}
}
\value{
lista di valori per metadato
}
\description{
Ritorna i valori dei metadati (o per singolo metadato)
}
