% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadati.r
\name{values_for}
\alias{values_for}
\title{Ritorna i valore del metadato per ogni singola serie}
\usage{
values_for(x, name, key)
}
\arguments{
\item{x}{istanza di grafo}

\item{name}{nomi di oggetti}

\item{key}{chiave del metadato}
}
\value{
lista di valori per metadato
}
\description{
Ritorna i valore del metadato per ogni singola serie
}
