% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{.buildConnection}
\alias{.buildConnection}
\title{Factory di connessioni al database Postgresql}
\usage{
.buildConnection(userid = whoami(), password = flypwd())
}
\arguments{
\item{userid}{userid utente}

\item{password}{password utente (defaults to flypwd)}
}
\description{
Factory di connessioni al database Postgresql
}
\note{
Funzione interna
}

