% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_dag.r
\name{assert_dag}
\alias{assert_dag}
\title{Controlla se il network e' un DAG}
\usage{
assert_dag(network)
}
\description{
TODO: dovrebbe tornare una lista di cicli se trovati. con la nuova versione di igraph non funziona.
}
