% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.r
\name{create_graph}
\alias{create_graph}
\title{crea ex-novo un istanza di grafo nel databae}
\usage{
create_graph(x, tag, con)

create_graph(x, tag, con, ...)
}
\arguments{
\item{x}{istanza di Grafo}

\item{tag}{identificativo della versione}

\item{con}{connessione al DB}

\item{`...`}{other optional params (like msg)}
}
\description{
crea ex-novo un istanza di grafo nel databae
}
