% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadati.r
\name{.deleteMeta}
\alias{.deleteMeta}
\title{elimina un metadato dal DBI}
\usage{
.deleteMeta(x, name, key, value)
}
\arguments{
\item{x}{istanza di grafo}

\item{name}{nome della serie da cui eliminare il metadato}

\item{key}{nome del metadato}

\item{value}{valore del metadato (se non specificato, rimuove tutti i  metadati
con la chiave specificata)}
}
\description{
elimina un metadato dal DBI
}
