% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{deleteMeta}
\alias{deleteMeta}
\title{Removes an attribute}
\usage{
deleteMeta(object, ts_name, attrName, attrValue)
}
\arguments{
\item{object}{graph instance}

\item{ts_name}{name of the timeseries}

\item{attrName}{name of the attribute}

\item{attrValue}{value of the attributes}
}
\description{
Removes an attribute
}
