% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{deleteMeta}
\alias{deleteMeta}
\title{Funzioni del package `grafo`}
\usage{
deleteMeta(graph, tsName, attrName, attrValue)
}
\description{
implementazione di listNodes di `package::grafo`
}
\seealso{
`grafo::deleteMeta`
}

