\docType{methods}
\name{deleteMeta}
\alias{deleteMeta}
\title{Funzioni del package `grafo`}
\usage{
deleteMeta(graph, tsName, attrName, attrValue)
}
\arguments{
  \item{object}{un oggetto di classe('grafo')}

  \item{tsName}{nome di una serie del grafo}

  \item{attrName}{nome del metadato da cancellare}

  \item{attrValue}{valore da cancellare, per i metadati con
  valori multipli}
}
\description{
implementazione di listNodes di `package::grafo`
}
\seealso{
`grafo::deleteMeta`
}

