% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{deleteMeta}
\alias{deleteMeta}
\title{Funzioni del package `grafo`}
\usage{
deleteMeta(graph, tsName, attrName, attrValue)
}
\description{
implementazione di listNodes di `package::grafo`
}
\seealso{
`grafo::deleteMeta`
}

