% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{deps}
\alias{deps}
\title{Ritorna le dipendeze di una serie}
\usage{
deps(x, name)
}
\arguments{
\item{x}{istanza di GrafoDB}

\item{name}{nome della serie}
}
\value{
un character array di nomi, NULL se la serie non ha dipendenze
}
\description{
Ritorna come array di caratteri i nomi delle serie su cui la serie specificata
dal parametro `name` dipende
}
