% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\docType{methods}
\name{describe}
\alias{describe}
\title{Funzioni del package `grafo`}
\usage{
describe(graph, nodes, order, mode, plot)
}
\description{
implementazione di describe di `package::grafo`
}
\seealso{
`grafo::describe`
}

