% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{disconnect}
\alias{disconnect}
\title{proxy the dbDisconnect based on `getenv` value}
\usage{
disconnect(con)

disconnect(con, env)
}
\arguments{
\item{con}{connection to disconnect}

\item{env}{environment to check against}
}
\description{
if `env` is `test`, is a no-op.
Otherwise delegates the call to `DBI::dbDisconnect`
}
