% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{downgrf}
\alias{downgrf}
\title{Ritorna i figli delle serie}
\usage{
downgrf(x, name)

downgrf(x, name, livello)
}
\arguments{
\item{x}{un istanza di GrafoDB}

\item{name}{array di nomi di serie}

\item{livello}{numero di livelli (ordine) da considerare (di default, tutti)}
}
\value{
nomi di serie
}
\description{
ritorna i nomi delle serie che sono generate dalle serie date in `name`
}
