% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{evaluate_impl}
\alias{evaluate_impl}
\title{Implementazione del generic `evaluate` definito nel package `grafo`
per la classe `GrafoDB`}
\usage{
evaluate_impl(object, v_start = NULL, ...)
}
\arguments{
\item{object}{GrafoDB instance}

\item{v_start}{node to be evaluated, if `NULL` evaluates all nodes}

\item{...}{eventual other params (like `debug`, for internal testing)}
}
\value{
il grafo con i dati correttamente valutato
}
\description{
Implementazione del generic `evaluate` definito nel package `grafo`
per la classe `GrafoDB`
}
