% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{.evaluate}
\alias{.evaluate}
\alias{.evaluate00}
\title{Implementazione del generic `evaluate` definito nel package `grafo`
per la classe `GrafoDB`}
\usage{
.evaluate(object)

.evaluate(object, v_start)
}
\value{
il grafo con i dati correttamente valutato
}
\description{
Implementazione del generic `evaluate` definito nel package `grafo`
per la classe `GrafoDB`
}
