% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{evaluate}
\alias{evaluate}
\alias{evaluate,GrafoDB-method}
\title{Implementazione del generic `evaluate` del package `grafo`
per la classe `GrafoDB`}
\usage{
evaluate(object)

evaluate(object, v_start)

\S4method{evaluate}{GrafoDB}(object, v_start = NULL, deep = F, ...)
}
\description{
Esegue il calcolo delle serie storiche presenti in questo Database
}
\note{
il generic e' definito nel package `grafo`
}
\examples{
\dontrun{
g <- GrafoDB(...) # istanzia il grafo
g <- evaluate(g)  # lo valuta tutto
v_start <- "ZERIQ"
g <- evaluate(g, v_start) # lo valuta solo a partire da ZERIQ
}
}
\seealso{
evaluate_impl
}
