% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.r
\name{.expr}
\alias{.expr}
\title{Ritorna le formule del GrafoDB}
\usage{
.expr(x, name)

.expr(x, name, echo=FALSE)
}
\arguments{
\item{x}{istanza di GrafoDB}

\item{nomi}{array di nomi di serie storiche}
}
\description{
Ritorna (o mostra su console, stdout) le formule delle serie `nomi`
}
