% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functions.r
\name{from.data.frame}
\alias{from.data.frame}
\title{converte un dataframe (caricato dal Database) in una timeseries `BIMETS`}
\usage{
from.data.frame(df)
}
\arguments{
\item{df}{data.frame compilato dal database}
}
\description{
converte un dataframe (caricato dal Database) in una timeseries `BIMETS`
}
\note{
i dati dal db sono memorizzati come stringhe JSON
}

