% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{from.data.frame}
\alias{from.data.frame}
\title{converte un dataframe (caricato dal Database) in una timeseries `ts`}
\usage{
from.data.frame(df)
}
\arguments{
\item{df}{data.frame compilato dal database}
}
\description{
converte un dataframe (caricato dal Database) in una timeseries `ts`
}
\note{
i dati dal db sono memorizzati come stringhe JSON
}
