% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.r
\name{getConflicts}
\alias{getConflicts}
\title{Ritorna i conflitti per un Grafo}
\usage{
getConflicts(x)

getConflicts(x, name)
}
\arguments{
\item{x}{istanza del GrafoDB}

\item{name}{vettore di nomi di serie}
}
\value{
un data.frame con le informazioni del conflitto
}
\description{
Ritorna i conflitti di un grafo, tutti per serie specificate dal
parametro `name`
}
