% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{getData}
\alias{getData}
\title{Funzioni del package `grafo`}
\usage{
getData(graph, tsNames)
}
\description{
implementazione di getData di `package::grafo`
}
\seealso{
`grafo::getData`
}

