% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{getData}
\alias{getData}
\alias{getData,GrafoDB,character-method}
\title{Funzioni del package `grafo`}
\usage{
getData(graph, ts_names)

\S4method{getData}{GrafoDB,character}(graph, ts_names)
}
\description{
implementazione di getData di `package::grafo`
}
\seealso{
`grafo::getData`
}
