% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{getdb}
\alias{getdb}
\title{Carica i dati dal DB}
\usage{
getdb(name, tag)
}
\arguments{
\item{name}{nome serie}

\item{tag}{id del grafo (default su `cf10`)}
}
\value{
una serie o una lista di serie
}
\description{
Carica i dati direttamente dal DB senza necessita' d'inizializzare
un `GrafoDB`
}
