% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getDependencies}
\alias{getDependencies}
\title{Funzioni del package `grafo`}
\usage{
getDependencies(graph, tsName)
}
\description{
implementazione di getDependencies di `package::grafo`
}
\seealso{
`grafo::getDependencies`
}

