% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{getDependencies}
\alias{getDependencies}
\title{Funzioni del package `grafo`}
\usage{
getDependencies(graph, tsName)
}
\description{
implementazione di getDependencies di `package::grafo`
}
\seealso{
`grafo::getDependencies`
}

