% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{getMetadata}
\alias{getMetadata}
\title{get Metadata for object}
\usage{
getMetadata(object, ts_name)

getMetadata(object, ts_name, full)
}
\arguments{
\item{object}{GrafoDB instance}

\item{ts_name}{object name}

\item{full}{boolean, triggers full report}
}
\description{
get Metadata for object
}
