% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlhelper.r
\name{sql_by_key}
\alias{sql_by_key}
\title{Factory di query (stringhe) SQL.}
\usage{
sql_by_key(x, key)

sql_by_key(x, key, param=value, param2, value2, ...)
}
\arguments{
\item{x}{istanza di SQLHelper}

\item{.key}{nome della query nel file INI}
}
\value{
un character array contenente la query SQL
}
\description{
Ritorna le query SQL cosi' come storicizzate nell'SQLHelper. Sostituisce
eventuali parametri (key=value) della query nell'ellipsis del metodo
}
\details{
Le query hanno una forma "select * from --tab-- where param='--value--'";
Il metodo si occupa di sostiture '--tab--' e '--value--' secondo i valori
passati come argomento: Es: `sql_by_key(helper, KEYQUERY, tab="A", value="B")`
risultera' in `select * from A where param='B'`
}
