% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{getTask}
\alias{getTask}
\title{Funzioni del package `grafo`}
\usage{
getTask(graph, tsName)
}
\description{
implementazione di getTask di `package::grafo`
}
\seealso{
`grafo::getTask`
}

