% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GrafoDB}
\alias{GrafoDB}
\alias{GrafoDB-class}
\title{Database Grafo}
\arguments{
\item{...}{don't know if used}
}
\description{
Classe per accedere ai dati immagazzinati in PostgreSQL del Grafo CF
}
\section{Slots}{

\describe{
\item{\code{data}}{\link[hash]{hash} containing changed data of this \link{GrafoDB}}

\item{\code{tag}}{edition of this Graph}

\item{\code{network}}{\link[igraph]{igraph} containing the DAG}

\item{\code{metadati}}{temporary metadata added by the user to be saved on the DB}
}}

