% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/core.r
\docType{class}
\name{GrafoDB}
\alias{GrafoDB}
\alias{GrafoDB-class}
\title{Database Grafo}
\arguments{
\item{...}{don't know if used}
}
\description{
Classe per accedere ai dati immagazzinati in PostgreSQL del Grafo CF
}
\section{Slots}{

\describe{
\item{\code{data}}{\link[hash]{hash} containing changed data of this \link{GrafoDB}}

\item{\code{tag}}{edition of this Graph}

\item{\code{network}}{\link[igraph]{igraph} containing the DAG}

\item{\code{metadati}}{temporary metadata added by the user to be saved on the DB}

\item{\code{ordinal}}{ordinale dei dati storici (0 per la produzione corrente)}

\item{\code{touched}}{serie modificate in area di lavoro}
}}
\examples{
\dontrun{
   g = GrafoDB("cf10") # istanzia il grafo chiamato 'cf10'
                       # in questo caso ordinal e' 0
   g = GrafoDB("cf10p2") # istanzia il grafo cf10 con provvisorio p2;
                         # in questo caso ordinal e' 2
}
}

