% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\docType{methods}
\name{initialize}
\alias{GrafoDB-initialize}
\alias{initialize}
\title{costruttore per la classe GrafoDB}
\usage{
\S4method{initialize}{GrafoDB}(.Object, tag = "cf10")
}
\description{
costruttore per la classe GrafoDB
}

