% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{initialize,GrafoDB-method}
\alias{initialize,GrafoDB-method}
\title{Default Constructor for GrafoDB}
\usage{
\S4method{initialize}{GrafoDB}(.Object, tag = "cf10")
}
\arguments{
\item{.Object}{GrafoDB object instance}

\item{tag}{label identifying the GrafoDB}
}
\description{
Default Constructor for GrafoDB
}
