% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{.init}
\alias{.init}
\title{Funzione d'inizializzazione del grafo.}
\usage{
.init(.Object, tag = "cf10")
}
\arguments{
\item{.Object}{(creato da new)}

\item{tag}{tag del grafo (default=`cf10`)}
}
\value{
un istanza di grafo popolata correttamente secono i parametri (`tag`)
}
\description{
Questa funzione va utilizzata nell'initialize S4 dell'oggetto `GrafoDB`
}
\note{
e' stata scorporata dall'initialize S4 per finalita' di debug
}
