% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{isAggregate}
\alias{isAggregate}
\title{Funzioni del package `grafo`}
\usage{
isAggregate(graph, tsName)
}
\description{
implementazione di isAggregate di `package::grafo`
}
\seealso{
`grafo::isAggregate`
}
