\docType{methods}
\name{isAggregate}
\alias{isAggregate}
\title{Funzioni del package `grafo`}
\usage{
isAggregate(graph, tsName)
}
\arguments{
  \item{graph}{un oggetto di classe('grafo')}

  \item{tsName}{nome della serie}
}
\description{
implementazione di isAggregate di `package::grafo`
}
\seealso{
`grafo::isAggregate`
}

