% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{isElementary}
\alias{isElementary}
\alias{isElementary,GrafoDB,character-method}
\title{Funzioni del package `grafo`}
\usage{
isElementary(graph, ts_name)

\S4method{isElementary}{GrafoDB,character}(graph, ts_name)
}
\description{
implementazione di isElementary di `package::grafo`
}
\note{
`GrafoDB` non prevede l'utilizzo di serie "elementari" come il `grafo`
      Quindi per compliance ritorna sempre `FALSE`, ma il metodo non ha senso
}
\seealso{
`grafo::isElementary`
}
