% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{isElementary}
\alias{isElementary}
\title{Funzioni del package `grafo`}
\usage{
isElementary(graph, tsName)
}
\description{
implementazione di isElementary di `package::grafo`
}
\note{
`GrafoDB` non prevede l'utilizzo di serie "elementari" come il `grafo`
      Quindi per compliance ritorna sempre `FALSE`, ma il metodo non ha senso
}
\seealso{
`grafo::isElementary`
}

