% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{isNode}
\alias{isNode}
\alias{isNode,GrafoDB,character-method}
\title{Funzioni del package `grafo`}
\usage{
isNode(graph, ts_name)

\S4method{isNode}{GrafoDB,character}(graph, ts_name)
}
\arguments{
\item{graph}{GrafoDB instance}

\item{ts_name}{name of the object}
}
\description{
implementazione di isNode di `package::grafo`
}
\seealso{
`grafo::isNode`
}
