% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{isNode}
\alias{isNode}
\title{Funzioni del package `grafo`}
\usage{
isNode(graph, tsName)
}
\description{
implementazione di isNode di `package::grafo`
}
\seealso{
`grafo::isNode`
}

