\docType{methods}
\name{isNode}
\alias{isNode}
\title{Funzioni del package `grafo`}
\usage{
isNode(graph, tsName)
}
\arguments{
  \item{graph}{un oggetto di classe('grafo')}

  \item{tsName}{nome della serie}
}
\description{
implementazione di isNode di `package::grafo`
}
\seealso{
`grafo::isNode`
}

