% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{isPrimitive}
\alias{isPrimitive}
\title{Funzioni del package `grafo`}
\usage{
isPrimitive(graph, tsName)
}
\description{
implementazione di isPrimitive di `package::grafo`
}
\seealso{
`grafo::isPrimitive`
}

