% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{isRoot}
\alias{isRoot}
\title{predicato che ritorna `TRUE` se le serie sono radici}
\usage{
isRoot(x, name)
}
\arguments{
\item{x}{istanza di GrafoDB}

\item{name}{vettore di nomi di serie}
}
\value{
`TRUE` se `name` sono serie senza archi entranti
}
\description{
Ritorna `TRUE` se le serie date nel parametro `name` non hanno serie
entranti
}
