% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadati.r
\name{keys_impl}
\alias{keys_impl}
\title{Ritorna le chiavi dei metadati}
\usage{
keys_impl(x)
}
\arguments{
\item{x}{istanza di grafo}
}
\value{
ritorna un dataframe con le chiavi dei metadati
}
\description{
Ritorna le chiavi dei metadati
}
