% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{leaves}
\alias{leaves}
\alias{leaves,GrafoDB-method}
\title{Returns all the leaves in a graph}
\usage{
leaves(x, ...)

\S4method{leaves}{GrafoDB}(x)
}
\arguments{
\item{x}{GrafoDB instance}

\item{...}{other params (unused)}
}
\value{
character array with the leaves names
}
\description{
Returns all the leaves in a graph
}
