% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{listAggregates}
\alias{listAggregates}
\alias{listAggregates,GrafoDB-method}
\title{Funzioni del package `grafo`}
\usage{
listAggregates(graph)

\S4method{listAggregates}{GrafoDB}(graph)
}
\arguments{
\item{graph}{GrafoDB instance}
}
\description{
implementazione di listAggregates di `package::grafo`
}
\seealso{
`grafo::listAggregates`
}
