% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{listElementaries}
\alias{listElementaries}
\title{Funzioni del package `grafo`}
\usage{
listElementaries(graph)
}
\description{
implementazione di listElementaries di `package::grafo`
}
\seealso{
`grafo::listElementaries`
}

