% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{listElementaries}
\alias{listElementaries}
\title{Funzioni del package `grafo`}
\usage{
listElementaries(graph)
}
\description{
implementazione di listElementaries di `package::grafo`
}
\seealso{
`grafo::listElementaries`
}

