\docType{methods}
\name{listElementaries}
\alias{listElementaries}
\title{Funzioni del package `grafo`}
\usage{
listElementaries(graph)
}
\arguments{
  \item{graph}{un oggetto di classe('grafo')}
}
\description{
implementazione di listElementaries di `package::grafo`
}
\seealso{
`grafo::listElementaries`
}

