% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{listNodes}
\alias{listNodes}
\title{Funzioni del package `grafo`}
\usage{
listNodes(graph)
}
\description{
implementazione di listNodes di `package::grafo`
}
\seealso{
`grafo::listNodes`
}
