% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{listPrimitives}
\alias{listPrimitives}
\alias{listPrimitives,GrafoDB-method}
\title{Funzioni del package `grafo`}
\usage{
listPrimitives(graph)

\S4method{listPrimitives}{GrafoDB}(graph)
}
\arguments{
\item{graph}{GrafoDB instance}
}
\description{
implementazione di listPrimitives di `package::grafo`
}
\seealso{
`grafo::listPrimitives`
}
