% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\docType{methods}
\name{listPrimitives}
\alias{listPrimitives}
\title{Funzioni del package `grafo`}
\usage{
listPrimitives(graph)
}
\description{
implementazione di listPrimitives di `package::grafo`
}
\seealso{
`grafo::listPrimitives`
}

