% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{roots}
\alias{roots}
\title{Ritorna le radici del grafo}
\usage{
roots(x)
}
\arguments{
\item{x}{istanza di grafo}
}
\value{
la lista di nomi delle radici del grafo
}
\description{
Ritorna le radici del grafo
}
