% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{schema_from_env}
\alias{schema_from_env}
\title{Ritorna il nome del file contenente lo schema dei dati in base all' `env`}
\usage{
schema_from_env(env = getenv())
}
\arguments{
\item{env}{ambiente di riferimento (di default e' `getenv`)}
}
\value{
percorso al file contenente lo schema dei dati
}
\description{
`env` ottenuto con `getenv` e' la variabile d'ambiente GRAFODB_ENV, di
default settata a `prod`. Negli ambienti di produzione e' necessario cambiarla
a `test` e modificare accordingly la configurazione
}
\note{
attualmente esistono solo due schemi, per SQLite (tipicamente `test`)
      e PostgreSQL (`prod`)
}
