\docType{methods}
\name{searchNode}
\alias{searchNode}
\title{Funzioni del package `grafo`}
\usage{
searchNode(graph, tsNames)
}
\arguments{
  \item{graph}{un istanza di grafo}

  \item{attrName}{chiave dell'attributo}

  \item{attrValue}{valore dell'attributo}
}
\description{
implementazione di searchNode di `package::grafo`
}
\seealso{
`grafo::searchNode`
}

