% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{searchNode}
\alias{searchNode}
\alias{searchNode,GrafoDB,character,character-method}
\title{Funzioni del package `grafo`}
\usage{
searchNode(graph, attr_name, attr_value)

\S4method{searchNode}{GrafoDB,character,character}(graph, attr_name, attr_value)
}
\arguments{
\item{graph}{instanza di GrafoDB}

\item{attr_name}{nome dell'attributo da cercare}

\item{attr_value}{valore da ricercare}
}
\description{
implementazione di searchNode di `package::grafo`
}
\seealso{
`grafo::searchNode`
}
