% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_grafo_impl.r
\name{setMeta}
\alias{setMeta}
\title{Imposta un metadato per una particolare serie}
\usage{
setMeta(object, ts_name, attr_name, value)
}
\value{
il GrafoDB modificato
}
\description{
Imposta un metadato `attr_name`=`value` per la serie `ts_name`
Se la serie non esiste, va in errore.
}
\examples{
\dontrun{
g <- GrafoDB()
g["A"] <- TSERIES(...)
g <- setMeta(g, "A", "FONTE", "CODICE_FONTE")

searchNode(g, "FONTE", "CODICE_FONTE") # deve tornare il nome "A"
}
}
