\docType{methods}
\name{showInternalChanges}
\alias{showInternalChanges}
\title{Funzioni del package `grafo`}
\usage{
showInternalChanges(graph)
}
\arguments{
  \item{object}{L'istanza di grafo che si vuole
  controllare}
}
\description{
implementazione di showInternalChanges di `package::grafo`
}
\seealso{
`grafo::showInternalChanges`
}

