% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{showInternalChanges}
\alias{showInternalChanges}
\title{Funzioni del package `grafo`}
\usage{
showInternalChanges(graph)
}
\description{
implementazione di showInternalChanges di `package::grafo`
}
\seealso{
`grafo::showInternalChanges`
}

