% Generated by roxygen2 (4.0.2): do not edit by hand
\name{.subsetting}
\alias{.subsetting}
\title{implementazione per GrafoDB del subsetting '[<-'}
\usage{
.subsetting(x, i, j, ..., value)
}
\arguments{
\item{x}{istanza di `GrafoDB`}

\item{i}{un character}

\item{value}{valore settato}

\item{j}{un'altro character o `missing`}

\item{...}{ellipsis dio solo sa a cosa servono}
}
\description{
implementazione per GrafoDB del subsetting '[<-'
}
\note{
funzione interna
}

