% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.r
\name{.subsetting}
\alias{.subsetting}
\title{implementazione per GrafoDB del subsetting '[<-'}
\usage{
.subsetting(x, i, j, ..., value)
}
\arguments{
\item{x}{istanza di `GrafoDB`}

\item{i}{un character}

\item{value}{valore settato}
}
\description{
implementazione per GrafoDB del subsetting '[<-'
}
\note{
funzione interna
}
