% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to.data.frame}
\alias{to.data.frame}
\title{converte una timeseries `BIMETS` in un data.frame.
funzione utilizzata per convertire il dato in una forma accettabile dal DB}
\usage{
to.data.frame(x)
}
\arguments{
\item{x}{una timeseries `BIMETS`}

\item{name}{nome da dare alla timeseries}
}
\value{
una rappresentazione a data.frame della serie `x`
}
\description{
converte una timeseries `BIMETS` in un data.frame.
funzione utilizzata per convertire il dato in una forma accettabile dal DB
}
\note{
funzione interna
}

