% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrf.r
\name{upgrf}
\alias{upgrf}
\title{Ritorna i genitori delle serie}
\usage{
upgrf(x, name)

upgrf(x, name, livello)
}
\arguments{
\item{x}{un istanza di GrafoDB}

\item{name}{array di nomi di serie}

\item{livello}{numero di livelli (ordine) da considerare (di default, tutti)}
}
\value{
nomi di serie
}
\description{
ritorna i nomi delle serie che entrano nelle serie date in `name`
}
