% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadati.r
\name{.values}
\alias{.values}
\alias{.values_by_key}
\alias{values_for}
\title{Ritorna i valori dei metadati (o per singolo metadato)}
\usage{
.values(x)

.values(x, key)

.values(x, name, key)
}
\arguments{
\item{x}{istanza di grafo}

\item{key}{chiave del metadato}
}
\value{
lista di valori per metadato

lista di valori per metadato
}
\description{
Ritorna i valori dei metadati (o per singolo metadato)

Ritorna i valore del lmetadato per ogni singola serie
}
